user/pyproject.toml
[tool.poetry]
name = "read-data-service"
version = "0.1.0"
description = ""
authors = ["usmanashrf <usmanashraf1510@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.18"
pyjwt = "^2.8.0"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg = "^3.2.1"
pydantic = {extras = ["email"], version = "^2.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
